chore: move worker handlers into WorldWorker
